chore: CustomerIOPluginOptionsIOS structure revised for clarity (#247)
Reorganizing the plugin configuration option in a way that makes them bit more intuitive. This included moving top level flags to be grouped together under the push notification object. 

For the time being the previous flags are deprecated but will still work.

## Change Log
In the app.json and app.config.js all the following push notification flags under the `ios`

```json
{
  "ios": {
    "showPushAppInForeground": true,
    "autoTrackPushEvents": true,
    "handleDeeplinkInKilledState": true,
    "disableNotificationRegistration": true,
    "autoFetchDeviceToken": true
  }
}

```
Are move to be part of the `pushNotification` object instead
```json
{
  "ios": {
    "pushNotification": {
      "showPushAppInForeground": true,
      "autoTrackPushEvents": true,
      "handleDeeplinkInKilledState": true,
      "disableNotificationRegistration": true,
      "autoFetchDeviceToken": true
    }
  }
}

```